updated consul registration according to PR feedback
