+++ pkg/discovery/constants.go
@@ -5,6 +5,9 @@ const (
 	TAG_MANAGED_SERVICE = "managedMicroservice"
 	TAG_COMPONENT_ATTRIBUTES = "componentAttributes"
 	TAG_SERVICE_NAME = "name"
+	TAG_VERSION = "version"
+	TAG_BUILD_NUMBER = "buildNumber"
+	TAG_BUILD_DATE_TIME = "buildDateTime"
 
 	COMPONENT_ATTRIBUTE_DELIMITER = "~"
 	KEY_VALUE_SEPARATOR = ":"
